requirements.txt
annotated-types==0.7.0
anthropic
anthropic==0.34.1
anyio==4.4.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
filelock==3.15.4
flask
Flask==3.0.3
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth-httplib2==0.2.0
google-auth==2.34.0
google-generativeai
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio-status==1.62.3
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.0
openai
openai==1.43.0
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.4
pyasn1_modules==0.4.0
pyasn1==0.6.0
pydantic_core==2.20.1
pydantic==2.8.2
pyparsing==3.1.4
python-dotenv
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rsa==4.9
safetensors==0.4.4
sniffio==1.3.1
sympy==1.13.2
tokenizers==0.19.1
torch
torch==2.4.0
tqdm==4.66.5
transformers
transformers==4.44.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.4
youtube_transcript_api
youtube-transcript-api==0.6.2
